Add support for Rack2 and Rack3.
