+++ .github/workflows/test.yaml
@@ -10,9 +10,11 @@ env:
 
 jobs:
   test:
-    name: ${{matrix.ruby}} on ${{matrix.os}}
+    name: ${{matrix.ruby}} on ${{matrix.os}} ${{matrix.gemfile}}
     runs-on: ${{matrix.os}}-latest
     continue-on-error: ${{matrix.experimental}}
+    env:
+      BUNDLER_GEMFILE: ${{matrix.gemfile}}
     
     strategy:
       matrix:
@@ -24,6 +26,11 @@ jobs:
           - "2.7"
           - "3.0"
           - "3.1"
+        
+        gemfile:
+          - gems/rack-v1.rb
+          - gems/rack-v2.rb
+          - gems/rack-v3.rb
         
         experimental: [false]
