+++ gems/rack-v1.rb
@@ -1,0 +1,3 @@
+eval_gemfile("../gems.rb")
+
+gem "rack", "~> 1.0"
