+++ gems/rack-v3.rb
@@ -1,0 +1,3 @@
+eval_gemfile("../gems.rb")
+
+gem "rack", "~> 3.0.0.beta1"
