+++ lib/protocol/rack/adapter.rb
@@ -20,146 +20,18 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-require 'console'
-
-require_relative 'constants'
-require_relative 'input'
-require_relative 'response'
+require_relative 'adapter/rack2'
+require_relative 'adapter/rack3'
 
 module Protocol
 	module Rack
-		class Adapter
-			PROTOCOL_HTTP_REQUEST = "protocol.http.request"
-		
-			# Initialize the rack adaptor middleware.
-			# @parameter app [Object] The rack middleware.
-			def initialize(app)
-				@app = app
-				
-				raise ArgumentError, "App must be callable!" unless @app.respond_to?(:call)
-			end
-			
-			def logger
-				Console.logger
-			end
-
-			# Unwrap raw HTTP headers into the CGI-style expected by Rack middleware.
-			#
-			# Rack separates multiple headers with the same key, into a single field with multiple lines.
-			#
-			# @parameter headers [Protocol::HTTP::Headers] The raw HTTP request headers.
-			# @parameter env [Hash] The rack request `env`.
-			def unwrap_headers(headers, env)
-				headers.each do |key, value|
-					http_key = "HTTP_#{key.upcase.tr('-', '_')}"
-					
-					if current_value = env[http_key]
-						env[http_key] = "#{current_value};#{value}"
-					else
-						env[http_key] = value
-					end
+		module Adapter
+			def self.new(app)
+				if ::Rack::RELEASE >= "3"
+					Adapter::Rack3.wrap(app)
+				else
+					Adapter::Rack2.wrap(app)
 				end
-			end
-			
-			# Process the incoming request into a valid rack `env`.
-			#
-			# - Set the `env['CONTENT_TYPE']` and `env['CONTENT_LENGTH']` based on the incoming request body. 
-			# - Set the `env['HTTP_HOST']` header to the request authority.
-			# - Set the `env['HTTP_X_FORWARDED_PROTO']` header to the request scheme.
-			# - Set `env['REMOTE_ADDR']` to the request remote adress.
-			#
-			# @parameter request [Protocol::HTTP::Request] The incoming request.
-			# @parameter env [Hash] The rack `env`.
-			def unwrap_request(request, env)
-				if content_type = request.headers.delete('content-type')
-					env[CGI::CONTENT_TYPE] = content_type
-				end
-				
-				# In some situations we don't know the content length, e.g. when using chunked encoding, or when decompressing the body.
-				if body = request.body and length = body.length
-					env[CGI::CONTENT_LENGTH] = length.to_s
-				end
-				
-				self.unwrap_headers(request.headers, env)
-				
-				# HTTP/2 prefers `:authority` over `host`, so we do this for backwards compatibility.
-				env[CGI::HTTP_HOST] ||= request.authority
-							
-				if request.respond_to?(:remote_address)
-					if remote_address = request.remote_address
-						env[CGI::REMOTE_ADDR] = remote_address.ip_address if remote_address.ip?
-					end
-				end
-			end
-			
-			def make_env(request)
-				request_path, query_string = request.path.split('?', 2)
-				server_name, server_port = (request.authority || '').split(':', 2)
-				
-				env = {
-					PROTOCOL_HTTP_REQUEST => request,
-					
-					RACK_INPUT => Input.new(request.body),
-					RACK_ERRORS => $stderr,
-					RACK_LOGGER => self.logger,
-
-					# The request protocol, either from the upgrade header or the HTTP/2 pseudo header of the same name.
-					RACK_PROTOCOL => request.protocol,
-					
-					# The HTTP request method, such as “GET” or “POST”. This cannot ever be an empty string, and so is always required.
-					CGI::REQUEST_METHOD => request.method,
-					
-					# The initial portion of the request URL's “path” that corresponds to the application object, so that the application knows its virtual “location”. This may be an empty string, if the application corresponds to the “root” of the server.
-					CGI::SCRIPT_NAME => '',
-					
-					# The remainder of the request URL's “path”, designating the virtual “location” of the request's target within the application. This may be an empty string, if the request URL targets the application root and does not have a trailing slash. This value may be percent-encoded when originating from a URL.
-					CGI::PATH_INFO => request_path,
-					CGI::REQUEST_PATH => request_path,
-					CGI::REQUEST_URI => request.path,
-
-					# The portion of the request URL that follows the ?, if any. May be empty, but is always required!
-					CGI::QUERY_STRING => query_string || '',
-					
-					# The server protocol (e.g. HTTP/1.1):
-					CGI::SERVER_PROTOCOL => request.version,
-					
-					# The request scheme:
-					RACK_URL_SCHEME => request.scheme,
-					
-					# I'm not sure what sane defaults should be here:
-					CGI::SERVER_NAME => server_name,
-					CGI::SERVER_PORT => server_port,
-				}
-				
-				self.unwrap_request(request, env)
-				
-				return env
-			end
-			
-			# Build a rack `env` from the incoming request and apply it to the rack middleware.
-			#
-			# @parameter request [Protocol::HTTP::Request] The incoming request.
-			def call(request)
-				env = self.make_env(request)
-				
-				status, headers, body = @app.call(env)
-				
-				return Response.wrap(status, headers, body, request)
-			rescue => exception
-				Console.logger.error(self) {exception}
-				
-				body&.close if body.respond_to?(:close)
-
-				return failure_response(exception)
-			end
-			
-			private
-
-			# Generate a suitable response for the given exception.
-			# @parameter exception [Exception]
-			# @returns [Protocol::HTTP::Response]
-			def failure_response(exception)
-				Protocol::HTTP::Response.for_exception(exception)
 			end
 		end
 	end
